[DF-10238] x64 bug fix. If role < 0, than roleId has incorrect value

+++ src/core/lib/core_qt_common/models/qt_item_model.cpp
@@ -239,7 +239,7 @@ QVariant QtItemModel::data( const QModelIndex &index, int role ) const
 		return QVariant();
 	}
 	
-	size_t roleId = role;
+	size_t roleId = static_cast<unsigned int>(role);
 	switch (role)
 	{
 	case Qt::DisplayRole:
